Move `BestChain` and `ChainStore` to chainstore.rs
`BestChain` is defined at the very end of the crowded chain_state.rs, and it seems a type that we would want to put next to `DiskBlockHeader` (the two storage Floresta types).

Also perhaps it's a good idea to move the `ChainStore` trait definition to the file that has its name. Probably makes sense as in pruned_utreexo/mod.rs we define the two chain backend traits, which are higher level.
